moved paths and updated links
