+++ aries-site/src/examples/components/notifications/index.js
@@ -1,1 +1,2 @@
-export * from './notifications';
+export * from './StateVsStatusDiagram';
+export * from './Notification';
