+++ aries-site/src/examples/components/toast-notifications/index.js
@@ -1,3 +1,2 @@
-export * from './ToastNotification';
 export * from './ToastNotificationExample';
 export * from './ToastDiagram';
