+++ aries-site/src/pages/components/toast-notifications.mdx
@@ -8,8 +8,8 @@ import { Box } from 'grommet';
 <Example
   docs="https://v2.grommet.io"
   code={[
-      "https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/components/visualizations/ToastNotificationExample.js",
-      "https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/components/visualizations/ToastNotification.js",
+      "https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/components/toast-notifications/ToastNotificationExample.js",
+      "https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/components/notifications/Notification.js",
   ]}
   figma="https://www.figma.com"
 >
